weather_impact_analysis/requirements.txt
google-adk
opentelemetry-instrumentation-google-genai
pandas
numpy
python-dotenv
requests
